Add method to get last commit ID
Added get_last_commit_id method to git_utils to retrieve the ID of the last commit in the current repo. This will be used to show the commit ID in the UI before creating a new commit. Also updated app.py to display the last commit ID and clear the panel when there are no staged changes.